Download all tokens

+++ get_images.py
@@ -1,5 +1,18 @@
 import json, requests, shutil, os, sys, getopt
 
+def do_request(id, take, skip):
+    url = "https://api.fxhash.xyz/graphql/"
+    query = f"""{{
+        generativeTokensByIds(ids: {id}){{
+        objkts(take: {take}, skip: {skip}) {{
+              metadata
+              }}
+        }}
+    }}"""
+
+    return requests.post(url, json={'query': query})
+    
+
 def main(argv):
     GT_id = ''
     folder = ''
@@ -19,41 +32,48 @@ def main(argv):
     if not os.path.exists(folder):
         os.mkdir(folder)
 
-    url = "https://api.fxhash.xyz/graphql/"
-    query = f"""{{
-        generativeTokensByIds(ids: {GT_id}){{
-        objkts {{
-              metadata
-              }}
-        }}
-    }}"""
+    valid_request = True
+    token_count = 0
+    take_count = 20
 
-    r = requests.post(url, json={'query': query})
+    while valid_request:
+        # Do multiple requests to get all tokens, like a pagination.
+        r = do_request(GT_id, take_count, token_count)
+        token_count += take_count
 
-    if r.status_code == 200:
-        print("Request succesful")
-        binary = r.content
-        output = json.loads(binary)
-        output = output['data']['generativeTokensByIds'][0]['objkts']
+        if r.status_code == 200:
+            binary = r.content
+            output = json.loads(binary)
 
-        for objkt in output:
-            objkt_name = objkt['metadata']['name']
-            image_url = f"https://gateway.fxhash.xyz/ipfs/{objkt['metadata']['displayUri'][7:]}"
-            filename = f'{folder}/{objkt_name}.png'
+            output = output['data']['generativeTokensByIds'][0]['objkts']
 
-            r_img = requests.get(image_url, stream = True)
-            if r_img.status_code == 200:
-                r_img.raw.decode_content = True
+            for objkt in output:
+                objkt_name = objkt['metadata']['name']
+                image_url = f"https://gateway.fxhash.xyz/ipfs/{objkt['metadata']['displayUri'][7:]}"
+                filename = f'{folder}/{objkt_name}.png'
 
-                with open(filename,'wb') as f:
-                    shutil.copyfileobj(r_img.raw, f)
+                if os.path.exists(filename):
+                    print(f'{objkt_name} already downloaded')
+                    continue
 
-                print(f'{objkt_name} downloaded')
+                r_img = requests.get(image_url, stream = True)
+                if r_img.status_code == 200:
+                    r_img.raw.decode_content = True
 
-            else:
-                print(f'Error downloading {objkt_name}')
+                    with open(filename,'wb') as f:
+                        shutil.copyfileobj(r_img.raw, f)
 
-    else:
-        print(f'Error {str(r.status_code)}')
+                    print(f'{objkt_name} downloaded')
+
+                else:
+                    print(f'Error downloading {objkt_name}')
+                    valid_request = False
+
+            if len(output) < take_count:
+                break
+
+        else:
+            print(f'Error {str(r.status_code)}')
+            valid_request = False
 
 main(sys.argv[1:])
